Refactor Google Maps integration codebase
