+++ app/src/androidTest/java/com/github/geohunt/app/maps/GoogleMapClusteringTest.kt
@@ -8,52 +8,41 @@ import androidx.compose.ui.test.junit4.createComposeRule
 import com.github.geohunt.app.maps.marker.Marker
 import com.google.android.gms.maps.model.CameraPosition
 import com.google.android.gms.maps.model.LatLng
+import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import java.time.LocalDateTime
 import java.time.Month
 
-class GoogleMapViewTest {
+class GoogleMapClusteringTest {
     @get:Rule
     val composeTestRule = createComposeRule()
     private val epflCoordinates = LatLng(46.51958, 6.56398)
-    private val mockBitmap: Bitmap = Bitmap.createBitmap(100, 100, Bitmap.Config.ARGB_8888)
+    private val mockTestBitmap: Bitmap = Bitmap.createBitmap(100, 100, Bitmap.Config.ARGB_8888)
+    private var mockTestChallengeDatabase : List<Marker> = mutableListOf()
 
-    private val mockChallengeDatabase : List<Marker> = listOf<Marker>(
-        Marker(
-            LatLng(46.51958, 6.56398),
-            "Event 1",
-            "Expires on 1 May 2024 at 19:39",
-            mockBitmap,
-            LocalDateTime.of(2024, Month.MAY, 1, 19, 39, 12)
-        ),
-        Marker(
-            LatLng(46.519, 6.563),
-            "Event 2",
-            "Expires on 1 May 2024 at 19:39",
-            mockBitmap,
-            LocalDateTime.of(2024, Month.MAY, 1, 19, 39, 12)
-        ),
-        Marker(
-            LatLng(46.5192, 6.5632),
-            "Event 3",
-            "Expires on 1 May 2024 at 19:39",
-            mockBitmap,
-            LocalDateTime.of(2024, Month.MAY, 1, 19, 39, 12)
-        ),
-        Marker(
-            LatLng(46.5193, 6.5633),
-            "Event 4",
-            "Expires on 1 May 2024 at 19:39",
-            mockBitmap,
-            LocalDateTime.of(2024, Month.MAY, 1, 19, 39, 12)
-        ),
-    )
+    @Before
+    fun initMockChallengeDatabase() {
+        val mockChallengeDatabase =  mutableListOf<Marker>()
 
+        for (i in 1..100) {
+            mockChallengeDatabase.add(Marker(
+                LatLng(46.51958 + i * 0.01, 6.56398 + i * 0.01),
+                "Event $i",
+                "Expires on 1 May 2024 at 19:39",
+                mockTestBitmap,
+                LocalDateTime.of(2024, Month.MAY, 1, 19, 39, 12))
+            )
+        }
+
+        mockTestChallengeDatabase = mockChallengeDatabase
+    }
 
     fun init(cameraPosition: CameraPosition = CameraPosition(epflCoordinates, 9f, 0f, 0f)) {
         composeTestRule.setContent {
-            GoogleMapView(
+            loadChallenges(mockTestChallengeDatabase)
+
+            GoogleMapDisplay(
                 Modifier.testTag("Maps"),
                 cameraPosition = cameraPosition,
             )
@@ -61,29 +50,21 @@ class GoogleMapViewTest {
     }
 
     @Test
-    fun testMapIsDisplayed() {
-        init()
-        composeTestRule.onNodeWithTag("Maps")
-            .assertExists()
-    }
+    fun testClusteringAtZoomLevel8() {
+        init(CameraPosition(epflCoordinates, 8f, 0f, 0f))
 
-
-    @Test
-    fun testClickOnFirstThreeMockMarkersAtZoomLevel10() {
-        init(CameraPosition(epflCoordinates, 10f, 0f, 0f))
-        composeTestRule.onNodeWithContentDescription("Marker Icon for Event 1")
+        composeTestRule.onNodeWithTag("Cluster for 10 markers")
             .assertExists()
             .assertIsDisplayed()
 
-        composeTestRule.onNodeWithContentDescription("Marker Icon for Event 2")
+        composeTestRule.onAllNodesWithTag("Cluster for 19 markers")
+            .assertCountEquals(4)[0]
             .assertExists()
             .assertIsDisplayed()
 
-        composeTestRule.onNodeWithContentDescription("Marker Icon for Event 3")
+        composeTestRule.onNodeWithTag("Cluster for 14 markers")
             .assertExists()
             .assertIsDisplayed()
-            .assertHasClickAction()
-            .performClick()
     }
 
     @Test
